Install failed. Could not find addon with name: ember-cli-string-utils
Setting up mirage, a11y audits, and auth are going to be common to every
acceptance test so I think it makes sense to bake them into this
blueprint.

The downside of this approach is that we’ll have to keep an eye on [the
upstream blueprint] and merge in any important changes over time. I think
this is a reasonable trade-off to save us having to look up how to do
these bits of setup every single time.

[the upstream blueprint]: https://github.com/emberjs/ember.js/tree/master/blueprints/acceptance-test